Split search panel searches into their own feature
