+++ bee-redux/src/features/hintPanels/index.ts
@@ -2,3 +2,4 @@ export * from "./api/hintPanelsApiSlice";
 export * from "./api/hintPanelsSlice";
 export * from "./components/HintPanel";
 export * from "./components/HintPanels";
+export * from "./types";
