+++ bee-redux/src/features/searchPanelSearches/components/Result.tsx
@@ -1,10 +1,10 @@
-import { SearchResultProps } from "./Results";
 import { WordLengths } from "./WordLengths";
 import { WordCount } from "./WordCount";
 import { LettersPresent } from "./LettersPresent";
 import { ResultHeader } from "./ResultHeader";
 import { SubstringHintOutputKeys } from "@/features/hints";
-import { ResultKey } from "@/features/hints/components/searchPanel/ResultKey";
+import { ResultKey } from "@/features/searchPanelSearches/components/ResultKey";
+import { SearchResultProps } from "@/features/searchPanelSearches/types";
 
 export function Result({ resultData, statusTracking }: SearchResultProps) {
   const content = () => {
