+++ bee-redux/src/features/searchPanelSearches/index.ts
@@ -1,0 +1,3 @@
+export * from "./api/searchPanelSearchesApiSlice";
+export * from "./components/Result";
+export * from "./types";
