+++ bee-redux/src/features/searchPanelSearches/types/index.ts
@@ -1,0 +1,19 @@
+import {
+  GridRow,
+  SearchPanelSearchData,
+  SubstringHintDataCell,
+} from "@/features/hints";
+import { StatusTrackingKeys } from "@/features/hintPanels/types";
+
+export interface ResultData {
+  searchObject: SearchPanelSearchData;
+  results: GridRow;
+  total: SubstringHintDataCell;
+  excludedAnswers: number;
+}
+
+export interface SearchResultProps {
+  panelId?: number;
+  resultData: ResultData;
+  statusTracking: StatusTrackingKeys;
+}
